+++ src/app/globals.css
@@ -1,4 +1,4 @@
-@tailwind base;
+/* @tailwind base;
 @tailwind components;
 @tailwind utilities;
 
@@ -31,4 +31,4 @@
         @apply text-sm text-gray-600 dark:text-gray-400
     }
 
-}
+} */
