+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,11 @@
+import { Box, Stack } from '@mui/material'
+
+export function Footer() {
+  return (
+    <Stack direction="column">
+      <Box>Footer 1</Box>
+      <Box>Footer 2</Box>
+      <Box>Footer 3</Box>
+    </Stack>
+  )
+}
