+++ src/components/Header/Logo.tsx
@@ -1,0 +1,11 @@
+import { Box, Stack } from '@mui/material'
+
+export function Logo() {
+  return (
+    <Stack direction="column">
+      <Box>Logo 1</Box>
+      <Box>Logo 2</Box>
+      <Box>Logo 3</Box>
+    </Stack>
+  )
+}
